Move API to the separate file.
